+++ src/main/java/hattrickdata/League.java
@@ -1,0 +1,7 @@
+package hattrickdata;
+
+import lombok.Builder;
+
+@Builder
+public record League(int id, String name) {
+}